# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The responses of the open api in swagger focus on the data field rather than the code and the message fields",
        "title": "Rainbow-APP-Service",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "console.nftrainbow.cn",
    "basePath": "/apps",
    "paths": {
        "/bot/invite_url": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get invite url",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "get invite url",
                "operationId": "GetInviteUrl",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "socialTool",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/bot/server": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get bot servers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "get bot servers",
                "operationId": "GetBotServers",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "social_tool",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FindBotServersResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "insert bot server",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "insert bot server",
                "operationId": "InsertBotServer",
                "parameters": [
                    {
                        "description": "insert_bot_server_request",
                        "name": "insert_bot_server_request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.InsertBotServerReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BotServer"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/bot/server/activities": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get activites of user bot servers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "get activites of user bot servers",
                "operationId": "GetActivitiesOfUserBotServers",
                "parameters": [
                    {
                        "type": "string",
                        "name": "activityName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "contractAddress",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "socialTool",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.FindBotServerActivitiesResult"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/bot/server/authcode": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get bot server roles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "get bot server roles",
                "operationId": "GetAuthCode",
                "parameters": [
                    {
                        "type": "string",
                        "name": "server_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "social_tool",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginutils.CommonMessage"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/bot/server/channels": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get bot server channels",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "get bot server channels",
                "operationId": "GetBotServerChannels",
                "parameters": [
                    {
                        "type": "string",
                        "name": "serverId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "socialTool",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/services.Channel"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/bot/server/push/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "push notification to social tool server",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "push notification to social tool server",
                "operationId": "Push",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "push_info ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ginutils.CommonMessage"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/bot/server/pushinfo/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "update push info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "update push info",
                "operationId": "UpdatePushInfo",
                "parameters": [
                    {
                        "description": "update push info request",
                        "name": "push_info_req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.PushInfoReq"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "push_info ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PushInfo"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/bot/server/roles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get bot server roles",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "get bot server roles",
                "operationId": "GetBotServerRoles",
                "parameters": [
                    {
                        "type": "string",
                        "name": "serverId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "name": "socialTool",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/services.Role"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/bot/server/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get bot server",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "get bot server",
                "operationId": "GetBotServer",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.BotServer"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
        "/bot/server/{id}/pushinfo": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "add push info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Bot"
                ],
                "summary": "add push info",
                "operationId": "AddPushInfo",
                "parameters": [
                    {
                        "description": "add push info request",
                        "name": "push_info_req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.PushInfoReq"
                        }
                    },
                    {
                        "type": "integer",
                        "description": "bot_server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PushInfo"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    },
                    "500": {
                        "description": "Internal Server error",
                        "schema": {
                            "$ref": "#/definitions/appService_errors.RainbowAppServiceErrorDetailInfo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "appService_errors.RainbowAppServiceErrorDetailInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "ginutils.CommonMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.FindBotServerActivitiesResult": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.PlattenBotServerActivity"
                    }
                }
            }
        },
        "models.FindBotServersResult": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BotServer"
                    }
                }
            }
        },
...
        "services.InsertBotServerReq": {
            "type": "object",
            "required": [
                "auth_code",
                "server_id",
                "social_tool"
            ],
            "properties": {
                "auth_code": {
                    "type": "string"
                },
                "server_id": {
                    "type": "string"
                },
                "social_tool": {
                    "type": "integer"
                }
            }
        },
...
    }
}